# Backend/Agent/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: MNTRK by MoStar Industries AI Agent API
  description: |
    This API provides robust and accurate tools for tracking Mastomys Natalensis populations, analyzing ecological trends, and supporting Lassa fever outbreak management in Nigeria.
  contact:
    name: Mastomys Tracker Support Team
    url: https://www.mo-overlord.tech/
    email: akanimo@57vflx.onmicrosoft.com
  version: 1.4.0
servers:
  - url: https://www.mo-overlord.tech/
    description: Primary Production Server (Nigeria-focused)
paths:
  /test:
    post:
      summary: Test endpoint
      operationId: test_post
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          description: Internal server error
      x-openapi-router-controller: swagger_server.controllers.test_controller
...
  /chat:
    post:
      summary: Process a chat request
      operationId: process_chat_request
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                  description: The chat prompt
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                    description: The chat response
      x-openapi-router-controller: swagger_server.controllers.default_controller
...